roberto-ivr-processor/requirements.txt
dateparser==0.6.0
Django==1.11.29
djangorestframework-xml==1.3.0
djangorestframework==3.6.4
graypy==0.3
luhn==0.2.0
mysqlclient==1.3.13
xmltodict==0.12.0
zeep[xmlsec]==2.4.0
itg-django-utils==0.1.7
